network: use an ActivePeersRef instead of ActivePeers type
In order to prepare for supporting shutdown of the network while Network
handles remain un-dropped, introduce ActivePeersRef and use it in
Network in place of ActivePeers.

Using a weak pointer in Network handles will allow for the ActivePeers
type to be dropped once ConnectionManager finishes shutting down,
causing the broadcast sender to be dropped and subscribers properly
notified of the network shutting down.