wire: introduce a lightweight handshake
Introduce a lightweight handshake that is performed after the server
side of a connection is able to fully complete the mTLS process and
verify the client side's certificate. Adding this handshake eliminates
the client side from prematurely thinking that the connection has been
fully established.

Before this patch the process looked like:

1. Client initiates contact with the server
2. Server presents its TLS certificate to the client
3. Client verifies the server's certificate <-- Client thinks the connection is complete
4. Client presents its TLS certificate to the server
5. Server verifies the client's certificate <-- Server thinks the connection is complete

After this patch this looks like:

1. Client initiates contact with the server
2. Server presents its TLS certificate to the client
3. Client verifies the server's certificate
4. Client presents its TLS certificate to the server
5. Server verifies the client's certificate
6. Server sends ACK <-- Server thinks the connection is complete
7. Client recieves ACK <-- Client thinks the connection is complete